backend/requirements.txt
Django
djangorestframework
drf-nested-routers
django-filter
pillow
openpyxl
gunicorn
psycopg2-binary